updated programming page
