+++ web/src/components/channel_config/ChannelProgrammingConfig.tsx
@@ -1,14 +1,10 @@
-import { Box, Button, Paper, Stack } from '@mui/material';
+import { Box, Stack } from '@mui/material';
 import { DateTimePicker } from '@mui/x-date-pickers';
 import dayjs, { Dayjs } from 'dayjs';
 import { createContext, useState } from 'react';
 import { useSlideSchedule } from '../../hooks/programming_controls/useSlideSchedule.ts';
 import { usePreloadedChannelEdit } from '../../hooks/usePreloadedChannel.ts';
-import {
-  resetLineup,
-  setChannelStartTime,
-} from '../../store/channelEditor/actions.ts';
-import useStore from '../../store/index.ts';
+import { setChannelStartTime } from '../../store/channelEditor/actions.ts';
 import AddProgrammingButton from './AddProgrammingButton.tsx';
 import AdjustScheduleControls from './AdjustScheduleControls.tsx';
 import ChannelProgrammingList from './ChannelProgrammingList.tsx';
@@ -28,7 +24,6 @@ export const ScheduleControlsContext = createContext<ScheduleControlsType>({
 export function ChannelProgrammingConfig() {
   const { currentEntity: channel } = usePreloadedChannelEdit();
 
-  const programsDirty = useStore((s) => s.channelEditor.dirty.programs);
   const slideSchedule = useSlideSchedule();
 
   const handleStartTimeChange = (value: Dayjs | null) => {
@@ -56,60 +51,52 @@ export function ChannelProgrammingConfig() {
       }}
     >
       <Box display="flex" flexDirection="column">
-        <Paper sx={{ p: 2 }}>
-          <Box display="flex" justifyContent={'flex-start'}></Box>
+        <Box display="flex" justifyContent={'flex-start'}></Box>
 
-          <Stack
-            direction={{ xs: 'column', sm: 'row' }}
-            gap={{ xs: 1 }}
-            sx={{
-              display: 'flex',
-              pt: 1,
-              mb: 2,
-              columnGap: 1,
-              alignItems: 'center',
-            }}
-          >
-            <Box sx={{ mr: { sm: 2 }, flexGrow: 1 }}>
-              <DateTimePicker
-                label="Programming Start"
-                value={startTime}
-                onChange={(newDateTime) => handleStartTimeChange(newDateTime)}
-                slotProps={{ textField: { size: 'small' } }}
-              />
-            </Box>
-            {showScheduleControls && (
-              <Stack
-                direction={{ xs: 'column', sm: 'row' }}
-                sx={{
-                  display: 'flex',
-                  pt: 1,
-                  mb: 2,
-                  columnGap: 1,
-                  alignItems: 'center',
-                  justifyContent: 'flex-end',
-                  flexGrow: 1,
-                }}
-              >
-                <AdjustScheduleControls />
-              </Stack>
-            )}
-            <Button
-              variant="contained"
-              onClick={() => resetLineup()}
-              disabled={!programsDirty}
+        <Stack
+          direction={{ xs: 'column', sm: 'row' }}
+          gap={{ xs: 1 }}
+          sx={{
+            display: 'flex',
+            pt: 1,
+            mb: 2,
+            columnGap: 1,
+            alignItems: 'center',
+          }}
+        >
+          <Box sx={{ mr: { sm: 2 }, flexGrow: 1 }}>
+            <DateTimePicker
+              label="Programming Start"
+              value={startTime}
+              onChange={(newDateTime) => handleStartTimeChange(newDateTime)}
+              slotProps={{ textField: { size: 'small' } }}
+            />
+          </Box>
+          {showScheduleControls && (
+            <Stack
+              direction={{ xs: 'column', sm: 'row' }}
+              sx={{
+                display: 'flex',
+                pt: 1,
+                mb: 2,
+                columnGap: 1,
+                alignItems: 'center',
+                justifyContent: 'flex-end',
+                flexGrow: 1,
+              }}
             >
-              Reset
-            </Button>
-            <ChannelProgrammingTools />
-            <ChannelProgrammingSort />
-            <AddProgrammingButton />
-          </Stack>
+              <AdjustScheduleControls />
+            </Stack>
+          )}
 
-          <ChannelProgrammingList
-            virtualListProps={{ width: '100%', height: 600, itemSize: 35 }}
-          />
-        </Paper>
+          <ChannelProgrammingTools />
+          <ChannelProgrammingSort />
+          <AddProgrammingButton />
+        </Stack>
+
+        <ChannelProgrammingList
+          virtualListProps={{ width: '100%', height: 600, itemSize: 35 }}
+        />
       </Box>
     </ScheduleControlsContext.Provider>
   );
